requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
asgiref==3.8.1
    # via django
boto3==1.34.147
    # via -r requirements.in
botocore==1.34.147
    # via
    #   boto3
    #   s3transfer
build==1.2.1
    # via pip-tools
chardet==5.2.0
    # via reportlab
click==8.1.7
    # via pip-tools
cssselect2==0.7.0
    # via svglib
dj-database-url==2.2.0
    # via -r requirements.in
django==5.0.7
    # via
    #   dj-database-url
    #   django-classy-tags
    #   django-cms
    #   django-entangled
    #   django-filer
    #   django-formtools
    #   django-parler
    #   django-polymorphic
    #   django-sekizai
    #   django-storage-url
    #   django-storages
    #   django-treebeard
    #   djangocms-admin-style
    #   djangocms-alias
    #   djangocms-frontend
    #   djangocms-versioning
    #   easy-thumbnails
django-classy-tags==4.1.0
    # via
    #   django-cms
    #   django-sekizai
django-cms @ git+https://github.com/fsbraun/django-cms@feat/menu-perf
    # via
    #   -r requirements.in
    #   djangocms-alias
    #   djangocms-attributes-field
    #   djangocms-frontend
    #   djangocms-text-ckeditor
    #   djangocms-versioning
django-entangled==0.5.4
    # via djangocms-frontend
django-filer==3.1.4
    # via
    #   -r requirements.in
    #   djangocms-frontend
django-formtools==2.5.1
    # via django-cms
django-fsm==2.8.2
    # via djangocms-versioning
django-parler==2.3
    # via djangocms-alias
django-polymorphic==3.1.0
    # via django-filer
django-sekizai==4.1.0
    # via django-cms
django-storage-url==0.12.0
    # via -r requirements.in
django-storages==1.14.4
    # via django-storage-url
django-treebeard==4.7.1
    # via django-cms
djangocms-admin-style==3.3.1
    # via
    #   -r requirements.in
    #   django-cms
djangocms-alias==2.0.1
    # via -r requirements.in
djangocms-attributes-field==3.0.0
    # via djangocms-frontend
djangocms-frontend==1.3.3
    # via -r requirements.in
djangocms-text-ckeditor==5.1.5
    # via -r requirements.in
djangocms-versioning==2.0.2
    # via -r requirements.in
easy-thumbnails[svg]==2.8.5
    # via
    #   -r requirements.in
    #   django-filer
    #   djangocms-frontend
furl==2.1.3
    # via django-storage-url
html5lib==1.1
    # via djangocms-text-ckeditor
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
lxml==5.2.2
    # via svglib
orderedmultidict==1.0.1
    # via furl
packaging==24.1
    # via
    #   build
    #   django-cms
    #   djangocms-text-ckeditor
pillow==10.4.0
    # via
    #   djangocms-text-ckeditor
    #   easy-thumbnails
    #   reportlab
pip-tools==7.4.1
    # via -r requirements.in
psycopg2==2.9.9
    # via -r requirements.in
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
python-dateutil==2.9.0.post0
    # via botocore
reportlab==4.2.2
    # via
    #   easy-thumbnails
    #   svglib
s3transfer==0.10.2
    # via boto3
six==1.16.0
    # via
    #   furl
    #   html5lib
    #   orderedmultidict
    #   python-dateutil
sqlparse==0.5.1
    # via django
svglib==1.5.1
    # via easy-thumbnails
tinycss2==1.3.0
    # via
    #   cssselect2
    #   svglib
typing-extensions==4.12.2
    # via dj-database-url
urllib3==2.2.2
    # via botocore
uwsgi==2.0.26
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wheel==0.43.0
    # via pip-tools
whitenoise==6.7.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
